refactor: moving token helper functions to libs
